+++ src/main/java/com/travel/booking/domain/odsay/city/CityInfoService.java
@@ -1,0 +1,11 @@
+package com.travel.booking.domain.odsay.city;
+
+import com.travel.booking.domain.odsay.bus.BusTerminalDTO;
+
+import java.util.List;
+
+public interface CityInfoService {
+    List<CityDTO> getCityList();
+
+
+}
